move wizard types to root level
