+++ src/components/core/wizard/desktop-wizard/index.ts
@@ -1,2 +1,1 @@
 export { default as DesktopWizard } from './desktop-wizard';
-export type { DesktopWizardProps } from './desktop-wizard';
